Renmaes `useSignTransaction` to `useSignAndSendTransaction`
