+++ src/hooks/public/index.ts
@@ -8,6 +8,7 @@ export * from 'hooks/public/use-list-item';
 export * from 'hooks/public/use-sign-out';
 export * from 'hooks/public/use-sign-out';
 export * from 'hooks/public/use-sign-transaction';
+export * from 'hooks/public/use-sign-and-send-transaction';
 export * from 'hooks/public/use-sign-message';
 export * from 'hooks/public/use-user-wallet';
 export * from 'hooks/public/use-user';
